requirements.txt
fastapi
uvicorn
requests
pyyaml
python-dotenv
pydantic
